comparison_pipeline/requirements.txt
pytest
pylint
python-dotenv
psycopg2
pandas
beautifulsoup4
requests
lxml
boto3
botocore
